cocoa,TopicRecordsTable: sort records by timestamp, sort differentially
Sorting by timestamp provides a much better experience when a topic
has many partitions, especially when those partitions are weighted
differently (eg. when  one partition is far ahead compared to the
others).  However, sorting by timestamp usually means the data has to
be sorted _more_ (i.e. more iterations are required), which makes
sorting 2-3x slower when there are many records.  This change
compromises by only sorting newly-added records when `appending`.
This way the minimal amount of data that has to be sorted will be, but
there is a chance the end result won't be perfectly sorted.  Mostly,
that should be hardly noticeable and the performance difference (in
extreme cases as much as 500x) is definitely worth it.